+++ client/src/components/pages/home/index.js
@@ -1,7 +1,7 @@
 import { state } from '../../../init/state.js';
 import { instructions } from './instructions.js';
 import { numberOfQu } from './number-of-questions.js';
-import { startQuizButton } from './startButton.js';
+import { startQuizButton } from './start-button.js';
 
 /**
  * The home page.
