+++ client/src/components/pages/quiz/quiz-question-panel.js
@@ -2,7 +2,7 @@ import { state } from '../../../init/state.js';
 
 import { divElement } from '../../shared/div.js';
 import { createQuizButtonsDiv } from './quiz-buttons.js';
-import { createRestartQuizButtonPanel } from './reStartButton.js';
+import { createRestartQuizButtonPanel } from './restart-button.js';
 
 export const quizQuestionPanel = () => {
   //Main display panel for the question
